requirements.txt
aiogram==2.25.2
environs~=8.0.0
Django==5.0.4
psycopg2-binary
pillow==10.3.0
django-modeltranslation==0.18.11
requests==2.31.0
